refactor(timeline): dedupe nest traversal, resolver snapshots, and nest validation
- EditorViewModel.timelineResolver() replaces hand-rolled timelinesById snapshots at 9 call sites
- Timeline.reachableTimelines() backs FCPXML collectNests, XMEML timecode cache, and cycle check
- Shared nestBlockReason()/hasAudioClips validation for agent and UI nest paths
- Clip.freshenIds() unifies group-id remapping (regenerateIds, decomposeNest)
- Nesting moves to EditorViewModel+Nesting.swift; shared nest export warnings helper
- AppTheme opacity constants in media tiles
