+++ Sources/PalmierPro/Generation/Submission/MusicGenerationSubmission.swift
@@ -34,11 +34,10 @@ struct MusicGenerationSubmission {
         var videoURL: String?
         if mode == .videoToMusic {
             onPhase(.exporting)
-            let timelinesById = Dictionary(uniqueKeysWithValues: editor.timelines.map { ($0.id, $0) })
             let mp4 = try await TimelineRenderer.render(
                 timeline: editor.timeline,
                 resolver: editor.mediaResolver,
-                resolveTimeline: { timelinesById[$0] },
+                resolveTimeline: editor.timelineResolver(),
                 missingMediaRefs: editor.missingMediaRefs,
                 startFrame: source.startFrame,
                 frameCount: source.frameCount,
