+++ Sources/PalmierPro/MediaPanel/MediaTab/FolderTileView.swift
@@ -27,7 +27,7 @@ struct FolderTileView: View {
                 .fill(Color(white: 1.0, opacity: AppTheme.Opacity.subtle))
             Image(systemName: "folder.fill")
                 .font(.system(size: AppTheme.FontSize.display, weight: AppTheme.FontWeight.light))
-                .foregroundStyle(AppTheme.Accent.primary.opacity(0.85))
+                .foregroundStyle(AppTheme.Accent.primary.opacity(AppTheme.Opacity.prominent))
             if childCount > 0 {
                 countBadge
             }
